ASD Update At-scale-debug version to 1.5.4
Important Notice:
Please note that if the ASD application base code is updated to
version 1.5.4 or later, it is mandatory to also update the JTAG
driver code to match the version in the OpenBMC Linux kernel dev-6.1.
We have recently introduced a new functionality in the JTAG driver to
support the RunTCK command. To accommodate this new feature, the ASD
application has undergone compatibility changes. These changes are
not backward compatible with previous versions of the JTAG driver and
may cause the ASD to fail with a "failed to initialize JTAG driver"
symptom.

Tested:
Confirm ASD version reported at connection is 1.5.4

Signed-off-by: Corona, Ernesto <ernesto.corona@intel.com>